don't cast to int just to widen back to a long (#3190)

+++ support/cas-server-support-jdbc-monitor/src/main/java/org/apereo/cas/monitor/JdbcDataSourceHealthIndicator.java
@@ -20,7 +20,7 @@ public class JdbcDataSourceHealthIndicator extends AbstractPoolHealthIndicator {
     private final JdbcTemplate jdbcTemplate;
     private final String validationQuery;
 
-    public JdbcDataSourceHealthIndicator(final int maxWait,
+    public JdbcDataSourceHealthIndicator(final long maxWait,
                                          final DataSource dataSource,
                                          final ExecutorService executor,
                                          final String validationQuery) {
